LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/inheritance-expected.txt


FAIL Property font-family inherits assert_not_equals: got disallowed value "\"Not Initial!\""
PASS Property font-feature-settings has initial value normal
FAIL Property font-feature-settings inherits assert_equals: expected "\"smcp\", \"swsh\" 2" but got "\"swsh\" 2, \"smcp\""
PASS Property font-kerning has initial value auto
PASS Property font-kerning inherits
FAIL Property font-language-override has initial value normal assert_true: font-language-override doesn't seem to be supported in the computed style expected true got false
FAIL Property font-language-override inherits assert_true: font-language-override doesn't seem to be supported in the computed style expected true got false
PASS Property font-optical-sizing has initial value auto
PASS Property font-optical-sizing inherits
PASS Property font-size has initial value 16px
PASS Property font-size inherits
FAIL Property font-size-adjust has initial value none assert_true: font-size-adjust doesn't seem to be supported in the computed style expected true got false
FAIL Property font-size-adjust inherits assert_true: font-size-adjust doesn't seem to be supported in the computed style expected true got false
FAIL Property font-stretch has initial value 100% assert_equals: expected "100%" but got "normal"
FAIL Property font-stretch inherits assert_equals: expected "75%" but got "condensed"
PASS Property font-style has initial value normal
PASS Property font-style inherits
FAIL Property font-synthesis has initial value weight style assert_equals: expected "weight style" but got "style weight small-caps"
PASS Property font-synthesis inherits
PASS Property font-variant has initial value normal
FAIL Property font-variant inherits assert_equals: expected "none" but got "no-common-ligatures no-discretionary-ligatures no-historical-ligatures no-contextual"
PASS Property font-variant-alternates has initial value normal
PASS Property font-variant-alternates inherits
PASS Property font-variant-caps has initial value normal
PASS Property font-variant-caps inherits
PASS Property font-variant-east-asian has initial value normal
PASS Property font-variant-east-asian inherits
PASS Property font-variant-ligatures has initial value normal
PASS Property font-variant-ligatures inherits
PASS Property font-variant-numeric has initial value normal
PASS Property font-variant-numeric inherits
PASS Property font-variant-position has initial value normal
PASS Property font-variant-position inherits
PASS Property font-variation-settings has initial value normal
FAIL Property font-variation-settings inherits assert_not_equals: got disallowed value "\"wght\" 700"
FAIL Property font-weight has initial value 400 assert_equals: expected "400" but got "normal"
PASS Property font-weight inherits
